[#004] Обновил основной экран
Привел к общему дизайну
Добавил кнопку - Рассчитать
Добавил кнопку - Удалить
Добавил отображение алерта при 1 входе на экран
При нажатии на кнопку рассчитать данные сохраняются в БД
